+++ core/src/main/scala/io/fintrospect/parameters/Value.scala
@@ -1,0 +1,6 @@
+package io.fintrospect.parameters
+
+trait Value[T] extends Any {
+  self: AnyVal with Product =>
+  val value: T
+}